+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.0.3
+
+Added new permissions found in the AWS Policy Generator.
+Added new autoscaling permissions not yet in the AWS Policy Generator.
+
 ## 0.0.2
 
 - Bug fix. The minimizer would throw an exception when given a large value for minchars as it was using the python [-1] syntax to retrieve the last element of a list.  This has been replaced with the dart .last attribute.